Fix Clutz aliasing for TypeScript 3.7.
In TypeScript 3.7, renamed exports no longer resolve in nested
namespaces: https://github.com/microsoft/TypeScript/issues/34612. That
change is working as intended, so this PR updates tsickle's alias emit
to handle the situation.

The remaining problem is that TS3.6 and before reject `export {foo}`, as
foo resolves to the local name foo in the module and thus this export
creates a circular definition. The workaround is to indirect through the
mangled name.