drop extra 'exports = {}' emit
We added 'exports = {}' to our emit to allow top-level mutable exports,
but it turns out to have a million surprising interactions with other
JSCompiler assumptions.

The net effect of this on user code is:

1) JS code that imports TS code used to require two-level destructuring like:
  const x = goog.require('y');
  const {a, b} = goog.require('y');
and now can just be written like:
  const {a, b} = goog.require('y');

2) Mutating top-level exports is illegal.
  export let x = 3;
  export function f() { x = 4; }
This was actually already broken under both TS and JSCompiler in some
circumstances, but now it fails to compile.  In our experience so far
it's not hard to refactor such code to instead look like:
  let x = 3;
  export function getX() { return x(); }
  export function f() { x = 4; }

3) Using 'namespace' breaks even harder.  Namespaces never really worked
under tsickle, but now they fail to compile.  One workaround is to
tranform
  export namespace x { ... }
into
  namespace x { ... }
  export {x};

The ultimate workaround for any breakages is to add
  // @ts-ignore
  exports = {};
to opt back in to the old behavior.