Handle `{default as Foo}` imports from Clutz.
In ES modules (and thus TypeScript), the default export is simply a
property named `default` on the exports object. tsickle however maps
default imports from `goog:` modules directly to the exports object, to
better match common `goog.module` usage.

Previously, tsickle failed to recognize `{default as Foo}` style
renaming, causing it to alias an incorrect type symbol to the export, in
turn causing failed type resolution in Closure (see the test for an
example). This change recognizes the renaming style, fixing the problem.