Refactor & fix rest parameter handling.
This change mostly refactors rest parameter handling to extract some of
the complex logic from the main parameter handling loop.

In addition, it fixes a subtle bug where tsickle would, if it failed to
resolve the array member type of a rest parameter, just continue
emitting the type as is. This "fail open" behaviour can cause issues,
such as emitting `Array<Array<?>>` for a non-representable rest tuple
type, instead of `Array<?>`.

Not a huge improvement and clearly a corner case, but I believe this
overall improves correctness and legibility of the code.

Fixes #1120.