Allow @ExportsDecoratedItems to be applied to protected and private items.
The items will be private and protected to other TypeScript code, but will still be emitted as @export in closure output.

This is a compromise, but an improvement over the current situation, where fields that are logically protected or private are public to both TypeScript and Closure Compiler. It does carry the downside that TS code authors may be overly confident in their privacy, but this seems a worthy trade. protected and private are just guidelines after all.

Internal global presubmit passed. Details at cl/282008467