Rationalize type/value conflict handling.
While separate in TypeScript, in Closure Compiler the type and value
namespaces are merged. This means tsickle must take care not to emit
duplicate symbols for type/value conflicts, and chooses the value over
the type.

That in turn means tsickle cannot generally emit references to types
that have a type value conflict, i.e. symbols in the TS API that have
both the Type and the Value SymbolFlag set.

However we special case two situations: (1) builtin symbols such as
`Array`, where we rely on the Closure definition and thus assume the
type/value conflict is handled on that level; and (2) symbols generated
by Clutz, which are also defined in Closure land so the conflict is not
a problem.

However previously we had two different code paths both handling parts
of this, one for type references in `@implements`, one for all other
type references. Both paths only handled either (1) or (2). This change
rationalizes the situation, handling (1) and (2) in both code paths.