Add the @ExportDecoratedItemsIfPublic decorator modifier.
@ExportDecoratedItemsIfPublic is a new decorator modifier for fields which should only be exported if public. When private or protected they are renamable and may be eliminated as dead code. This is more optimal version of @ExportDecoratedItems for fields which can only be accessed by reflection if they're public.

Our use case ExportDecoratedItemsIfPublic is the LitElement @property decorator, where public properties are accessible as HTML attributes and may be written by raw HTML files or various template languages (lit-html, soy, angular templates, etc). This obviously is taking place outside of the class hierarchy of the element and so is only valid for public properties, but it is still useful for private properties to participate in the update lifecycle that comes along with @property.

This change, combined with the already-landed https://github.com/angular/tsickle/pull/1125 we'll be able to remove the final @ExportDecoratedItems annotation from the LitElement decorators.