don't dealias types in heritage clauses
If you have
```
class Foo implements Bar
```
where `Bar` is some sort of alias, previously we would resolve
Bar to the underlying alias and emit that name instead in the
`@implements` bit.  This change makes us instead use the original
Bar.

You can see in the two affected tests that we have tests for the ways
in which you might expect this to fail -- for example, the test shows
that JSCompiler handles it properly when you implements a type alias.

PiperOrigin-RevId: 313779842