move tsickle's clutz pass to be a transformer
I want to add yet more Clutz-specific logic, so I split it to
its own file, and changed it into a transformer.  The intent is
the output has equivalent behavior to the before code, but the
specific syntax used is slightly different just for simplicity
of codegen.

This lets us delete our hook for writeFile during emit, because
we now transform AST first, rather than jamming in text after.

PiperOrigin-RevId: 325134638