Use typeTranslator for @extends/@implements blocks
When we have some code like

```
class A implements B {...}
```

We want to emit some JSDoc like

```
/**
 * @implements {B}
 */
```

tsickle did so by trying to grab and emit the named symbol in the
'heritage clause' (aka 'extends'/'implements').

But this is wrong for lots of more obscure cases.  For example, if
you `implements` a generic type, we need the generic args to show
up in the JSDoc.

Fix this by switching to using the full TypeTranslator machinery
for heritage.  It turns out that a lot of the special casing (of
e.g. type/value conflicts) that was done in heritage handling is
already implicitly handled by TypeTranslator, so this deletes a
bunch of code.

One downside of this approach is that using types means we necessarily
expand aliases.  For example, if `B` in the above was some reexport of
some other type `C`, then by using TypeTranslator we directly emit `C`
in the output.  This is vaguely bad because it violates strict deps, but
it's also maybe inevitable if you want to get all of the harder cases
right.  For example, imagine there was a default for the generic, like
`type B<T=string> = C<T>`.  In JS we need to insert which generic we
chose.

PiperOrigin-RevId: 326023374