simplify further the e2e clutz test
Separate from the golden testing, tsickle has a test that just feeds
**/*.d.ts into the TypeScript compiler, to verify that the generated
d.ts are valid.

Simplify this code to just glob for those files (rather than gathering
them from each test separately), so that we can move one file in particular
out of the goldens tree.

Moving this file out unblocks an upcoming change, which is more cautious
about the handling of d.ts files and goldens.  This particular moved .d.ts
is not a golden, it's just needed for this compilation test.

PiperOrigin-RevId: 327506169