import clutz references by path in tsickle output
When a TypeScript file references a JS target via Clutz, it will
either have an import of the `goog:` namespace or a direct reference
to the look of disapproval.

As part of a larger project to enable `resolve: true` in TypeScript,
which will make tsc follow imports, we need Clutz references to also
be backed by real path-based imports, so that the file that defines
a symbol or `goog:` namespace can be found by the compiler by first
following the path-based imports.  To do so, when generating a
`.d.ts` from TypeScript code, we transform its references to Clutz
to also include a path-based reference.

Testing this is a little subtle: our test suite wants to type-check
its inputs and generate d.ts, but in the case of Clutz the inputs
also include Clutz-generated d.ts.  So change the rules of tests to
say "a d.ts is an input only if there is no corresponding ts".  This
allows the new test to include both d.ts inputs (simulating outputs
from clutz) as well as d.ts outputs (goldens capturing what tsickle
generates from those inputs).

PiperOrigin-RevId: 327661284