fix the tsickle-clutz integration in the presence of inference
When TS causes a Clutz symbol to appear in a d.ts due to inferring
a type, the resulting AST no longer has parent nodes nor is it visible
to the type checker.

To fix, change the logic that finds Clutz references to work top-down
(not using node.parent) and also use some internal TS API for the symbol.

All of this would become unnecessary if we removed support for look of
disapproval in TS, but that is a much larger project.

PiperOrigin-RevId: 329961639