Fix infinite recursion bug in tsickle
Tsickle crashes on types that look like:

```
type A = B<A>;
```

When trying to translate the type `A` will resolve into `B<A>`, which will attempt a translation again, and so on until the stack limit is reached. This CL adds logic to bail out early if a type parameter is the equal to the type being translated.

PiperOrigin-RevId: 338153840