Fix aliases for imported types always point to actual decl spot
The main change is to make jsdoc types that are exported somewhere
always point to the actual place of definition instead of any of
intermediate `export *`s.

The necessity of this change is that with the previous commit, tsickle
generates `{(typeof ns1|typeof ns2)}` type annotation for certain
namespaces in externs. There is no problem with resolving values on such
namespaces, but it appears that the closure compiler does not resolve
types on such namespaces. Therefore, when a namespace to be used in type
annotation is to be determined, it always needs to be a namespace on
which that type is actually defined.

This contract has been abided for `.ts` transformations (fortunately),
but wasn't for externs generation. Therefore it changes `addImportAlias`
function to use the method used in `.ts` transformations.

Type --(Type.symbol)--> symbol --TypeTranslator#symbolToString()--> name