Add a test case for a recursive union type alias.
JSCompiler does not support recursive type aliases. tsickle protects against this by emitting recursive type references as `?`.

We received a report that this particular pattern crashes tsickle. I believe it was fixed already in 2ec19dd096223ab432b942eeb8c20abf92c7a748. tsickle ends up inlining the first level of the type (because it follows some aliases), and correctly emits the recursive case as `?`. This isn't great for JSCompiler, but seems a reasonable representation of the pattern.

PiperOrigin-RevId: 346496468