Fix optional chaining crash in Tsickle
The optional chaining transform was only correctly identifying property access with `?.`, but both `a?.[0]` and `a?.()` are also legal. This CL updates the transformation logic to exclude invalid nodes, rather than build up a list of working nodes (at this point we know we are in a property access chain).

PiperOrigin-RevId: 347082053