Side-effect imports of files with conflicting provides are supported.
This change adds a test and docs for intended behaviour: when TypeScript code
imports a JavaScript file in a side-effect import, there is no symbol to
import, and thus no conflict between multiple symbols in the file.

This also adds a bunch of docs, clarifies a comment on empty import blocks, and
adds more negative tests for various scenarios (empty import blocks, etc).

PiperOrigin-RevId: 350085379