During prodmode goog.module transformation, delete calls to ts.migrationExportsShim rather than rewriting them as module map insertions.
If the .closure.js file is being loaded into a browser (as part of a debug session), then the .tsmes.closure.js file probably is also. The .tsmes.closure.js file will create the necessary module map entry (because it contains a goog.module declaration). A map entry added by the .closure.js file is just a module loading error.

PiperOrigin-RevId: 368311499