use better workaround for quoted access
This tsickle code uses a private API within TypeScript, so unless
the access to it is quoted it gets property renamed.  But quoting
at the usage site ties the code to an unrenamed interface to that
API, which would then break in the other way if we ever managed to
compile the TS API into the same compilation as this client code.

Instead, the proper place to mark this distinction (whether this
function is renamed or not) lies in the place where we decide to
bundle TypeScript -- if it's bundled without renaming, we need an
extern to mark this as unrenameable, and otherwise, the dotted
access is fine.

PiperOrigin-RevId: 370758824