Repair the tsickle demo.
We forgot to publish an updated version that works after the latest API
changes.

This change moves demo to a new release that's compatible, adjusts its
TypeScript & tsickle version, and also includes its basic build in the
yarn test target (that, independently, we might want to break into
smaller tasks at some point).

It also fixes an incorrect argument in the release instructions.