add a test that shows ES modules can control their goog.module name
Normally in Clutz we generate modules via 'declare module' blocks.
But if Clutz generated module d.ts, we still need control over the
generated goog.module name that clients use.  This test verifies that
works for an ordinary ES module.

(Note that this only works in d.ts files.  If you use this pattern in
a .ts file, the module name passed to `goog.module(...)` currently ignores
the `__clutz_actual_namespace` attribute.  But we don't need this behavior
for .ts files anyway.)

PiperOrigin-RevId: 380215366