Ensure all variants of goog.tsMigration*ExportsShim are transformed/stripped.
When the shorthand variants were added, there was an oversight in stripping. This causes calls to be retained in the output files, which would crash at runtime.

This change also re-expresses tsmes support as a transformer; rather than a feature that creates shim files, with a side-effect of transforming tsmes calls, we now have a transformer for tsmes calls, which produces shim files as the side-effect. Restructuring in this way allows the transformation to share code with shim generation, which is useful since the transformation for development sources is very similar to the shim file content.

PiperOrigin-RevId: 381925633