Fix tsickle for TypeScript 4.4.
In TypeScript 4.4, the "hasTrailingComma" field is no longer writable. The fix is to use the ts.factory.createNodeArray API.

This change also removes several deprecation warnings by moving all the code to ts.factory. This is probably an incomplete fix though: the intended use of the factory API appears to get a handle to a factory from the transformator context.

PiperOrigin-RevId: 383112464