tsickle: special case array destructuring.
JSCompiler does not support tuple types, so tsickle cannot emit an appropriate type for JSCompiler. JSCompiler represents tuples using union types, `A|B`. This can cause de-optimizations, e.g. when accessing `value.someMethod` that should be disambiguated to `A.someMethod`, which then makes JSCompiler think there is a `B.someMethod`.

Without tuple type support, this cannot be fixed in the general case. However we can special case common patterns, such as destructuring out of a tuple type, e.g. in a variable statement or in a for/of loop.

    for (const [a_1, b_1] of x) {
      const a = /** @type {string} */ (a_1);
      const b = /** @type {number} */ (b_1);
    }

Now we can give precise types to a and b. This causes some code bloat during development, but the alias will be collapsed by JSCompiler in optimizations.

PiperOrigin-RevId: 383390260