Alias destructured tuple function parameters for more precise types.
This is a follow up to cr/383390260.

This also is a roll forward of the previously rolled back change. The fix in this change is that it account for automatic semicolon insertion when transforming an arrow function body into a return statement. Previously code that had line comments before the arrow function body would be broken.

Closure Compiler does not support tuple types, meaning tsickle emits a weak `?` type for tuple types. To improve type safety and optimization compatibility, we can however do better in specific situations:

    function f([a, b]: [string, number]) {}

Becomes:

    function f([a_1, b_1]) {
      /** @type {string} */
      const a = a_1;
      /** @type {number} */
      const b = b_1;
    }

I.e. the tuple type is still untyped, but use sites of the symbols get a precise type.

Interestingly, it does not seem generally useful to improve the tuple type, e.g. by representing `[A, B]` as `Array<A|B>`. The resulting type is imprecise and causes more deoptimizations/problems than benefits (slightly tighter types for JS users).

PiperOrigin-RevId: 388172356