tsickle: emit tsmes.js file in ES5 and Closure modes.
Previously, tsickle somewhat awkwardly emitted the tsmes shim only when running
in "closure" mode, which it probed by inspecting whether type annotation was
requested. In that mode, it'd emit `.tsmes.closure.js`, and also remove the
tsmes calls from source.

In ES5 mode, presumably to avoid emitting two files (?), it instead produced a
`.d.ts` file, and instead of replacing the tsmes calls, it'd inline equivalent
logic. This was presumably done so that in dev mode, things would still work at
runtime by registering the shim module ID.

This setup doesn't work though: various tools expect an actual goog.module()
call to be found for the shim module ID, so these do not work with the ES5
sources. In addition, emitting a different set of files depending on contextual
information, is confusing and breaks other tools that expect the same set of
files in either run mode.

This change:
* unconditionally always emits the .tsmes.js file, and leaves it to the
  compiler host to write to .closure.js or .js (same as we do for all other JS
  files).
* produces a .d.ts when the `.declaration` TS compiler option is set,
  equivalent to how .d.ts are produced from regular sources.
* remove Generator.run(), which previously performed very different operations
  depending on arguments. Instead, this inlines some of the logic into the
  transformer function and makes explicit calls into the generator's functions.
* refrains from further refactoring, though I'm leaving a TODO with some ideas.

PiperOrigin-RevId: 425621974