Improve test result legibility by setting a background color.
The symbols used to mark deletions and insertions are very subtle and hard to
see in big diffs. Adding a splash of background color makes the test much more
readable.

The drawback is that this will likely print poorly in some environments (e.g.
test results displayed in some browsers where the tool doesn't interpret
console colors), but that still seems preferable to just the tiny [ ] markers.

PiperOrigin-RevId: 440321079