Automatically add missing namespace names
Add missing namespace names in declaration merging namespaces.
Previously, any identifier that references a symbol in the
namespace had to be fully qualified so that the transformed
code properly compiled. This is no longer necessary. Instead of
generating an error when finding an unqualified reference, the
translator now automatically adds the missing qualifier.

PiperOrigin-RevId: 481671858