Properly transpile exported array destructuring.
Before, it was missing a variable declaration after traspilation for identifiers used in alias assignment expressions. The cause was missing exports on actual identifiers that needed import, and ended up getting removed by the passes coming aftewards, because jsdoc_transformer was renaming identifiers destructuring statements, but not the usage of those identifiers properly because exported symbol ends up being used as `exports.identifier_destructured`, but it was being used as `identifier_destructured` pattern.

PiperOrigin-RevId: 597281499